generative_ai/LLM_custom_inference_model_template/storage/templates/google-gemini/requirements.txt
google-auth
openai
requests
pandas